python/requirements.txt
grpcio-tools==1.46.1
google-auth==2.19.1
six==1.16.0
